+++ CHANGELOG.md
@@ -9,7 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Fixed
+
+### Removed
+
+## [0.44.1]
+
+### Added
+
+### Fixed
 * Map pin points to the wrongs location #APPS-990
+* Show Grab&Go placeholder if order list is empty #APPS-1282
 
 ### Removed
 * Removed leinweber customercard as payment method #APPS-883